+++ lib/tasks/db.rake
@@ -18,7 +18,7 @@ namespace :db do
   task create: [:environment] do
     puts "aact_proj db:  set search_path ..."
     con=ActiveRecord::Base.establish_connection(ENV['AACT_PROJ_DATABASE_URL']).connection
-    con.execute("alter role #{ENV['AACT_PROJ_DB_SUPER_USERNAME']} in database #{ENV['AACT_PROJ_DATABASE']} set search_path = public, #{Admin::Project.schema_name_list};")
+    con.execute("alter role #{ENV['AACT_PROJ_DB_SUPER_USERNAME']} in database #{ENV['AACT_PROJ_DATABASE']} set search_path = public, mesh_archive, #{Admin::Project.schema_name_list};")
     con.reset!
     Rake::Task["db:create"].invoke
   end
